+++ .gitignore
@@ -3,3 +3,10 @@ node_modules
 
 # vscode
 .vscode
+
+# envs
+.env
+
+# aws
+certs
+
