+++ src/server.js
@@ -1,9 +1,9 @@
 const http = require('http');
 const app = require('./app');
-const PORT = process.env.PORT || 3000
+const PORT = process.env.PORT || 3000;
 
 const server = http.createServer(app);
 
 server.listen(PORT, () => {
-    console.log('Server listening at port: ', PORT);
+  console.log('Server listening at port: ', PORT);
 });
